requirements.txt
streamlit
google-generativeai
pandas
sentence-transformers
faiss-cpu
